refactor: move the create version of process to process model
Refs: oph-1031
